winrt: Obtain new camera sample format and size on camera change
By Resetting d->cameraSampleformat to Format_User when the selected
video device is changed we make QWinRTCameraVideoRendererControl
reobtain the camera sample format and size during the next call of
queueBuffer. This is needed as the new camera might have other values
for these, which might result in a white screen after a camera change.

Task-number: QTBUG-67417
Change-Id: I29e28ddf4fc278aff00ad824bda61aedad91471d
Reviewed-by: VaL Doroshchuk <valentyn.doroshchuk@qt.io>