Fix video window parenting and positioning
The positioning code seemed to be expecting that the video window is
created as an application window but mm-renderer creates the window
as a child when a group is specified. When the application window
isn't located in the top-left corner of the display, this
assumption produces incorrect positioning of the video window.

Contrary to what a comment in the code indicated, the video
window does not need to be parented to the application window.
Doing so just makes it difficult to determine the correct position
for the video window.

Also ensure that the video window is resized when the metadata
information doesn't contain width/height. On QNX 7.0.0, the
code may fail to resize the window because the metadata doesn't
contain the width/height information under the expected key. A
floating point divide by zero occurs with undefined results.
I'll adjust for the width/height move in a separate commit.

Change-Id: I540c1798a03b7c03a4438d0852c80e4d086009be
Reviewed-by: Janne Koskinen <janne.p.koskinen@qt.io>
Reviewed-by: Rafael Roquetto <rafael@roquetto.com>