QSampleCache: Fix memory leak
and a warning on termination:
QCoreApplication::postEvent: Unexpected null receiver

During static deinit it is not possible to use deleteLater, as there is
no event loop running at this stage.

The crash that was described in QTBUG-30630 doesn't reproduce with
Qt 5.10, so I guess it should be safe to just delete.

Change-Id: I0c404d71e9f102018e314c890ff679f8c0952b07
Reviewed-by: VaL Doroshchuk <valentyn.doroshchuk@qt.io>