Android: Fix fetching video frame data from the GUI thread
If QAbstractVideoSurface is not used on rendering thread
added a way to pass shared opengl context via GLContext surface property.

Camera and media players that are created in Qt Quick will usually have
their rendering done on the render thread, but the callback for custom video
surface(s) (or probes) will usually be handle on the GUI thread, so to
make it convenient to for the user to access the frame data, which is rendered
on request, we now set-up context sharing and render to a fbo on the calling
thread, if needed.

Task-number: QTBUG-69333
Change-Id: I4cc5c1f741c82376f1402a047b946b59281c9a4c
Reviewed-by: Christian Strømme <christian.stromme@qt.io>