Automatically register the latest import version
This follows what was done in d28c9f6a for Qt Quick Controls 2.
The latest import version (e.g. 2.11 in Qt 5.11) will automatically be
registered whenever the Qt version is bumped.

This avoids needing to wait until a new type is added (or a new
revision is added to an existing type) before being able to use the
newest Qt Quick version.

Change-Id: Ib14052fdf2a4cfa7b71f739fe00ce9a95dadb0a0
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>