VideoOutput: Introduce flushMode property
Added flushMode property to QML VideoOutput element to define
what should be shown when flush is requested.

Takes affect only for QDeclarativeVideoRendererBackend
and when QSGVideoItemSurface is already started.

Flushing (passing empty video frame to the surface) is usually performed
when EndOfMedia or playback is stopped.
Which caused disappearing the content and blinking if playlist is used.

Using this property now possible to define
what frame (last, first or empty) should be shown when playback is stopped or finished.

By default shows empty frame (clears the video output).

To show a frame it requires to keep QVideoFrame and thus its data.

Task-number: QTBUG-37301
Task-number: QTBUG-49446
Change-Id: I3be5309217b9f543da804e3b616dee9d97fba65f
Reviewed-by: Andy Shaw <andy.shaw@qt.io>
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>