Windows: Repaint QVideoWidget when WM_PAINT or WM_ERASEBKGND received
Since updates are disabled, and no paint events will be sent to
QVideoWidget, need to somehow repaint the video window.

When BeginPaint is called outside of the expose event, the window might
not be refreshed. So moved calling the repainting within the expose event.

Also added repainting when the display rect is changed to immediate
refreshing.

Task-number: QTBUG-71326
Change-Id: I7c29b3f8ad35e33590c7b138a5141546ce3c1a42
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>
Reviewed-by: Christian Stromme <christian.stromme@qt.io>