Android: Don't postpone setting of camera's state
Currently if the application is inactive,
the state is postponed to be set and processed when the app becomes active.

This also postpones stateChanged() signal and saving the state value,
while it should be set (but not processed) immediately as done in another backends.

Added a fix to store the state and emit stateChanged regardless of
activity of the application.
When the application becames available, the state will be processed.

In case of an error while opening the camera:
- UnloadedState is emitted
- CameraError
- UnloadedStatus

Change-Id: Ie376d29366168a6b15eccea884f385070f94fe35
Task-number: QTBUG-73582
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>