Android: Fix status of QMediaRecorder after start/stop
If the camera is ready for capture -> QMediaRecorder::LoadedStatus.
If CaptureVideo is changed to CaptureStillImage -> StoppedState and
UnloadedStatus.
If camera's status is StoppingStatus -> StoppedState and UnloadedStatus.
If camera's status is LoadingStatus -> LoadingStatus.
If recording is requested -> RecordingState and RecordingStatus.

If recording is audio-only -> immediately LoadedStatus, before start and
after stop.

Fixes tst_QCameraBackend::testVideoRecording

Task-number: QTBUG-73582
Change-Id: I976c4e3afab529e0949571c4002e9ceba74cac97
Reviewed-by: Christian Strømme <christian.stromme@qt.io>