EVR: Remove setting of start/end times in QVideoFrame
Due to unknown bug in IMFTransform->ProcessOutput() when after seeking
sample times are not set correctly to provided IMFSample.

This causes current playback position to be not respected in video samples.
Even if you seeked to higher/lower position, sample time is always counted
from beginning.

Which makes no sense to use this feature because video frame's sample times
are not related to current position.

To test it, need just seek to arbitrary position and check how startTime
is not correlated to new position.

Fixes tst_QMediaPlayerBackend::seekPauseSeek

Task-number: QTBUG-65574
Change-Id: I897d75c055347cdcca38a84dc18f91800d070c09
Reviewed-by: Christian Strømme <christian.stromme@qt.io>