DirectShow: Don't hide returned error when video output was not provided
Since the video output could be set after doRender(), we do not report
an error in this case.

But need to report the error when the video output has not been submitted at all.
Otherwise it hides the error and not possible to determine playback issues.

Added a fix to remember an error while rendering without the video
output and handle it if playing is requested.

That means, if it is required, the video output must be set before play() is called.
Not all medias require valid video output, but if there is an error, we
need to report about it.

Fixes tst_QMediaPlayerBackend::playlist()

Task-number: QTBUG-65574
Change-Id: I9faae19c08ad0273545bb7617ea3a11539084f1f
Reviewed-by: Andy Shaw <andy.shaw@qt.io>