Gstreamer: Fix deadlock when state is requested in ASYNC mode
When new media is set to the player, updating of "show-preroll-frame" property from the video sink
is also requested:
    g_object_set(G_OBJECT(m_videoSink), "show-preroll-frame", value, NULL);

This produces emitting of a signal:
    g_signal_connect(G_OBJECT(sink), "notify::show-preroll-frame", G_CALLBACK(handleShowPrerollChange), sink);

Inside handleShowPrerollChange() the state of the video sink is requested with infinite timeout:
    gst_element_get_state(GST_ELEMENT(sink), &state, NULL, GST_CLOCK_TIME_NONE);

In case if the video sink performed an ASYNC state change,
means changing of the state is pending and need to wait,
this function will block up for infinite timeout.

But probably changing of the state is requested (and should be performed) on the same thread
where it is waiting for this change, it produces a deadlock.

Changing timeout to 10ms to avoid this block.

Fixes: QTBUG-72468
Change-Id: I06235ccfb8f76423f65ed192d5e8de6e60723e72
Reviewed-by: Christian Strømme <christian.stromme@qt.io>