Gstreamer: Pass GstUDPSrc's caps via "udpsrc.caps=" url's query item
Sometimes it is needed to configure udpsrc element to play a stream,
means to set some caps to GstUDPSrc element.
But currently there are no any ways to pass such caps.

Added parsing of the requested url to find "udpsrc.caps" query item
and to use it as the caps for udpsrc source element.
It allows to show streams by passing caps within url.

E.g. if the stream is created using

$ gst-launch-1.0 v4l2src ! videoconvert ! video/x-raw,format=I420,width=800,height=600 ! \
jpegenc ! rtpjpegpay ! udpsink host=127.0.0.1 port=5001

it could be shown via QMediaPlayer like:

MediaPlayer {
  source: "udp://127.0.0.1:5001/?udpsrc.caps=application/x-rtp,media=video,clock-rate=90000,encoding=JPEG,payload=26"
}

Change-Id: I6f9c20c6004a34bce5fd1d0073311b7c62a8010f
Reviewed-by: Christian Strømme <christian.stromme@qt.io>