GStreamer: Don't seek to the beginning when playing is requested
We promised to not provide any video frames before pause() or play() is called.
Regardless the media is loaded or not.

It should work like following:
setMedia(new)
// ...
// no video frames returned yet
pause()
// prerolled frame should be shown

To implement this, i.e. showing the prerolled frame after pause, we do seeking to the beginning.
But it is totally not necessarily when play() is requested,
since here is new media and playback will start from the beginning anyway.

This also produces a hang/stuck on some embedded systems with custom pipelines or playing from qrc.

So decided to show prerolled frame and this means to seek to the beginning only when pause() is called.

Task-number: QTBUG-65399
Change-Id: I9b3dc632fa0df4a1115d852c2d480fd5c7fd9ee5
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>