Fix crash when app is destroyed before QSample::load is finished
QSample objects live and handled in loading thread (which uses QNetworkAccessManager).
When the app is finished and going to be destroyed, all static objects are destroying as well.

In case if static QNetworkConfigurationManagerPrivate (which is used by QNetworkAccessManager)
is destroyed before static QSampleCache, and loading of the resource is not finished yet
(still executing QNetworkAccessManager::get()), this produces a crash.

Since the loading thread is started only when loading of new QSample is requested,
(and all events are also handled by this thread)
proposing a fix to wait before loading thread is finished when a sample is requested to be released.

This postpones deleting of the QSample either when new sample is requested to load
or when QSampleCache is destroyed.

This makes sure that no loading thread exists when all QSoundEffects objects
and afterwards QNetworkConfigurationManagerPrivate are already deleted.

Change-Id: I55669ea4c2796a48cae4f0465f7f74d89e393675
Fixes: QTBUG-76090
Reviewed-by: Christian Strømme <christian.stromme@qt.io>