Fix API of QCameraExposure, QCameraFocus, QCameraImageProcessing for Qt for Python
The classes inherit QObject, but make the destructor private, which
makes it impossible to add Qt for Python bindings.

Make destructors protected to fix this. This is arguably a BC break
for compilers which have the accessibility mangled in the symbol, but,
since the destructor was private, there cannot be any code using it.

Task-number: QTBUG-74422
Task-number: PYSIDE-1041
Change-Id: Id23d73aea8aedf15caafa20c048a80efd3e880bd
Reviewed-by: VaL Doroshchuk <valentyn.doroshchuk@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>