Move-enable QMediaTimeInterval
The class shouldn't have a user-provided copy ctor, because the compiler
could write it for us and the type would be trivial, but we can't remove
it in Qt 5, because that changes the way the class is passed by value
into functions (register instead of stack).

At least give it back its move special member functions, inhibited by the
user-provided copy ctor, and schedule the whole lot for removal in Qt 6.

Change-Id: Ic19ef43f3d313fea938aa769c067a736ec691337
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>