Eradicate the last Q_FOREACH and mark the module clean
This one is not trivial in that it isn't clear, a priori, that the
loop body doesn't re-enter the class and ends up modifying
m_videoProbes.

But MFVideoProbeControl::bufferProbed() just schedules a deferred call
to QMediaVideoProbeControl::videoFrameProbed() (via QueuedConnection),
so this cannot happen, because we end the loop before we re-enter the
event loop (if any) on this thread. Reentrancy from a different thread,
OTOH, is not possible because we're holding m_videoProbeMutex while
looping.

Change-Id: I6c1e22cd01112f0b2650e056f0b7f4a392370b84
Reviewed-by: VaL Doroshchuk <valentyn.doroshchuk@qt.io>