VideoOutput: Update geometry only when the surface is active
Since starting the surface could be done after updating the geometry of
VideoOutput element (paint event is received),
and this leads setting invalid geometry to renderer and never updates by correct one.
Because the geometry is not "dirty" anymore, means already set.

Need just to update geometry when (or keep trying before) the surface is already started with
proper video surface format.

Change-Id: I338d7fe355c20f2027c6231241714376e9b569b2
Fixes: QTBUG-76205
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>