GStreamer: Allow audio/x-raw codec in QAudioRecorder
Currently QAudioRecorder is implemented to find a codec and muxer that match some caps.

E.g. if we would want "audio/x-wav" container with PCM format, there is only one gst element
which supports it:

wavenc element accepts "audio/x-raw", "audio/x-alaw" or "audio/x-mulaw"
and provides buffers in "audio/x-wav" or "audio/x-rf64".

Next, we would need to find a codec which provides buffers in
"audio/x-raw", "audio/x-alaw" or "audio/x-mulaw". It might be mulawenc and alawenc.

And finally we combine them together to pipeline:
audio/x-raw -> mulawenc|alawenc -> wavenc -> audio/x-wav

And as a result there will be wav file and encoded in "audio/x-alaw" or "audio/x-mulaw"
and not in desired PCM format.

So suggesting to add audioconvert element to the audio encoders to allow to link
with various containers:

audio/x-raw -> audioconvert -> wavenc -> audio/x-wav

Change-Id: Icd77124267f5a0999625fa1fe4e3de3740218741
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
Reviewed-by: Andy Shaw <andy.shaw@qt.io>