DirectShow: Add startTime and endTime to QVideoFrame
If the application seeks to a new position, the Media Session restarts
the presentation clock at the specified seek time.
The media source might deliver samples with a slightly earlier time stamp
if the seek time does not fall on a key-frame boundary.

And IMFSample->GetSampleTime will return time that is related to
a position when the clock has been restarted, and can be also negative.

So if it needs to have times from the beginning, it would require to add seeked position
to times from IMFSample.

It also reverts c4de056a6aa44567cdbf2ce91a464e597ad4af8f

Fixes: QTBUG-77849
Change-Id: I1a7cb7bd18aee73087a61d2ed2c3d644ad0fbd50
Reviewed-by: Christian Strømme <christian.stromme@qt.io>