DirectShow: Round stop position down to available bytes in IAsyncReader
Currently requested IMediaSample might contain the end time that exceeds the
available bytes which causes returning an error in WaitForNext() and
stopping the playback.

Regarding to IAsyncReader::Request documentation:

The start and stop positions should match the alignment that was decided when the pins connected.
The stop position might exceed the real duration.
If so, the method rounds the stop position down to the actual alignment.

Fixes: QTBUG-77782
Change-Id: I644e25bfc6bb8f6d345b8424b79fb56490d82c0e
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>