Move QVideoFrame::Format_ABGR32 to the end of the Pixel Formats
Since ABGR32 was introduced in 5.13.0 but placed in the middle of the list,
which breaks enum values.

[ChangeLog] In Qt 5.13.0 binary compatibility was broken
for usage of the enum QVideoFrame::PixelFormat by introducing QVideoFrame::Format_ABGR32.
To minimize the impact of this, the break has been reverted,
thus introducing a binary compatibility break with earlier Qt 5.13.x versions,
but restoring compatibility with all earlier versions of Qt 5.

Change-Id: I3c490f2fc30f3b6914306b2d496724f12729338e
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>