Don't load the specialist videonode plugins when using webgl
When the webgl platform is being used then we do not want to load the
videonode plugins because this will cause it to potentially use a
backend that is not going to be supported by WebGL. This is due to the
fact that it will potentially use extensions to OpenGL (such as the
imx6 ones). But since the rendering is not happening on the device, but
in the browser then the standard backends will work fine in this case.

Fixes: QTBUG-79181
Change-Id: I03021d4ec01d9447dd1ca3815d32a167392ac781
Reviewed-by: Ville Voutilainen <ville.voutilainen@qt.io>