Android: Flush qrc to tmp file with original name
Since android.media.MediaPlayer does not support playing from data stream,
we create a temporary file and copy data to it.

Each such file contains unique name.

When the app is killed, dtor of QMediaPlayer is not called and thus
the temporary file is not deleted.

Also these files are located inside app's location, e.g.
/data/data/org.qtproject.example.app/files/, and no possibility to delete
them automatically.

Added a fix to rename the temporary file to original name which allows to keep
only one copy of the file when the app is killed.

Change-Id: I36c8bc4dd339fbf4091689a05f4d9c7febf046a5
Fixes: QTBUG-59517
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>