Make video surfaces constant
Video surface property for QVideoWidget, QGprahicsVideoItem and QDeclarativeVideoOutput
should be CONSTANT and kept the same value while the object is alive.

Note, currently this property can be suddenly changed, when videoSurface is mixed/used
together with source property.
To keep the patch simple, this will be fixed in next commits.

Change-Id: I80046524a70bcb70bb45e7f1bbae74bff1b3b0f2
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>