AVF: Fix applying viewfinder settings for iOS
1. If a resolution of the viewfinder is requested to be changed,
it is not needed to check available formats for highResolutionStillImageDimensions
(since it is viewfinder resolution and not still images).

format.highResolutionStillImageDimensions returns the highest resolution still image
that can be produced by the format.

Using this format will cause wrong resolution in the viewfinder.

2. Seems iOS requires to call beginConfiguration/commitConfiguration to apply any changes.
commitConfiguration is currently called when the camera gets active.
In case if settings are changed after starting the camera, they will not be applied.

To fix that, added beginConfiguration/commitConfiguration.

Fixes: QTBUG-79935
Change-Id: I60b6e08887a82e4337a2a302b8e5513c65b26e61
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>