AVF: Set playback audio session category while playing
Since when the application is active and then moved into the background
then audio can be played fine.
But if the application is already in the background then it will not play the audio.

Decided to set AVAudioSessionCategoryPlayback with AVAudioSessionCategoryOptionMixWithOthers
to audio session, which leads to restart playing from the background.

Fixes: QTBUG-69949
Change-Id: I18fb4904b0810b7dfb63585278edbb8ee2f02d05
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>