Introduce QAbstractVideoBuffer::MTLTextureHandle
Added MTLTextureHandle to render metal textures.
Is used by default if rhi is enabled for metal backend.

Also fixed the frame renderer to create new opengl context
and use provided one from the video surface as a share context.

To remember, when the quick item is created and updatePaintNode is called,
current gl context is set to the video surface as a property.
When the frame renderer is ready, it extracts the gl context and uses it as a share one.

Task-number: QTBUG-78678
Change-Id: I51ce666ca7c2adc10dd2c1d1dfed99cc9f596e2b
Reviewed-by: Christian Strømme <christian.stromme@qt.io>