AVFoundation: fix playback of files with spaces in their names
Used QUrl::toEncoded() instead of QUrl::toString().

Usage of QUrl::toString() leads to URLs which can contain spaces and as
result such URLs are not accepted by AVPlayer, and it remains in unknown
state because AVFMediaPlayerSession::processLoadStateChange() is not
called at all. And as result, QMediaPlayer waits media loading forever.

[ChangeLog][Platform Specific Changes][AVFoundation] Fixed playback of
files with spaces in their names.

Fixes: QTBUG-84585
Pick-to: 5.15
Change-Id: I69e522d048d88310803a7040f114ee808c18720b
Reviewed-by: VaL Doroshchuk <valentyn.doroshchuk@qt.io>