First step in cleaning up the plugin loading code
Stop assuming that there are multiple plugins providing
the same service. We'll have exactly one plugin for a
service in the future and not provide several.

The reason is that we will focus on having one good and
functional multimedia backend per platform, not the mix
of many semi-functional ones as we have today.

Change-Id: I809d4c99a56a48369d0ba283af78b1c41740d92a
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
Reviewed-by: Doris Verria <doris.verria@qt.io>