Integrate audio devices with WMF layer on Windows
The audio layer still uses the old Windows MM/DirectShow
APIs. Those APIs use a different way to identify audio
inputs and outputs (the WaveID). WMF uses a UUID as the
device identifier.

To enable interoperability, retrieve the WMF ID when
enumerating the devices. Use those as the device id()
in QAudioDeviceInfo, but also store the old style WaveID.

When creating a QAudioInput/Output, retrieve the WaveID
and use that one to create the audio device.

Change-Id: I02ffeb64b7da4bd321b543bc11278c37a2a115c6
Reviewed-by: Doris Verria <doris.verria@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>