Unify the media recorder controls for macOS and iOS
These days, macOS seems to have all the required functionality to
use the same code path as iOS. Since that code path is more flexible
try to use it on all platforms.
This changes the way we record audio/video to file in macOS
from using AVCaptureMovieFileOutput to using the AVAssetWriter
like in iOS.

*There are still some issues to be addressed:
- when recording audio via a bluetooth audio capture device,
the viewfinder freezes.
- a short lag when switching from viewfinder to mediarecorder
for video capturing
*Implement setting mute and volume while recording

Change-Id: I5654e117b4d9ba74b60c500ed109902e317248c4
Reviewed-by: Lars Knoll <lars.knoll@qt.io>