Remove some code that not needed anymore
qgstreamerimagecapture/encoder provides the same functionality.
And we won't be going with camerabin anyway as the API is still
unstable and doesn't give us the flexibility we need.

Some of the remaining code in here can be brought over to give
better camera support, although gst-photography seems to be
mostly unused.

Change-Id: Iae7e8c20545f108a53a1efd0d83fd917c520e9d8
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
Reviewed-by: Doris Verria <doris.verria@qt.io>