improve meta data support on Apple platforms
AVMetaDataItem is underdocumented (as most of the non trivial corners
of AVFoundation). The main issue is to retrieve data for meta data keys
that Apple does not consider 'common'. After tons of searching, I found
that the thing to use is the [item identifier] property. Those can then
be compared against a list of extended meta data identifiers to find a
mapping to our QMediaMetaData::Key enum.

The only place this was 'documented' was in some deprecated example code
from Apple that's not linked to by the regular documentation (search for
avmetadataeditor on Google).

Change-Id: Id70c24b1044012a4828201e7bb6f1bbba2ae2c36
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
Reviewed-by: Doris Verria <doris.verria@qt.io>