Add a QMediaEncoder class
This class is currently a simply a rename of QMediaRecorder, but
they will start to diverge in the next commits.

QMediaRecorder is now an all-in-one class, that contains
a QMediaCaptureSession and handles standard recording cases.

QMediaEncoder will in the next couple of commits loose APIs
that should be in other places in the capture pipeline.

Change-Id: Ied1098092920610dac08e966078d44a22110bcf7
Reviewed-by: Doris Verria <doris.verria@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>