Don't add cameraControl and imageCaptureControl on session initialization
Add API for adding and relasing a session's cameraControl and
imageCaptureControl.
The session is audio-only until attaching a camera to it, so initialize
the cameraControl on addCamera(). Do the same for the imageCaptureControl.
Still need to change for other platforms besides darwin.
To be followed by changes in the mediaEncoder too.

Change-Id: I69097c73b745ffef3727135312a36f1f1d2627e5
Reviewed-by: Lars Knoll <lars.knoll@qt.io>