Further refactor the video sink implementation on macOS/iOS.
Make sure we can deliver both SW buffers, textures and render onto
native surfaces.

As we need different implementations for the player and capture
pipelines, AVFVideoSink is now acting as a proxy to the class
doing the actual rendering.

Fold the windowed mode handling into the AVFVideoRenderer class and use
it when rendering windowed or fullscreen.

This also required some changes in AVFMediaPlayer to ensure that we always
have a AVPlayerLayer available because loading content happends asynchronously,
but we need the layer directly.

With this playback works both onto QVideoWidget and the QML VideoItem, although
there are some R<->B issues when rendering using software.

Change-Id: Ie6fe901f35989c5076eedaea4edf3b8c4d3f35f4
Reviewed-by: Doris Verria <doris.verria@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>