Remove QAbstractVideoFilter
The API covers a relatively rare use case, and is far from
ideal, as it doesn't cooperate very well with HW accerlerated
rendering.

Since QVideoFrame and the QML VideoOutput item offer the option to
get a texture for the frame, it's probably better to use that
as the basis for further processing.

Change-Id: I89deb915f660312fb3a83b066a5371f5454346da
Reviewed-by: Doris Verria <doris.verria@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>