Have the QPlatformMediaIntegration create the platformimageCapture
Follow the QPlatformCamera example and create the imageCapture control
by the QPlatformMediaIntegration instead of the capture session.
Add signals to the backend for when controls are added to the session.
Make image capture on macOS work again.

Adjust autotests and some code in QCameraImageCapture and move some
sanity checks from the backend into the cross platform code.

Change-Id: Iad596e9062cc39ec4c4db5af60ff6527867cbce2
Reviewed-by: Lars Knoll <lars.knoll@qt.io>